// src/common/enums/custom-messages.enum.ts
export const enum CUSTOM_MESSAGES {
  UNAUTHORIZED_TOKEN = 'Tài khoản hoặc mật khẩu không chính xác',
  INVALID_TOKEN = 'Token không hợp lệ',
  TOKEN_EXPIRED = 'Token hết hạn',
  FORBIDDEN = 'Bạn không có quyền truy cập vào API này',
  ERROR_MONGO_ID = 'ID phải là kiểu MongoId',
  CATEGORY_EXIST = 'Tên danh mục đã tồn tại',
  CATEGORY_ID_NOT_EXIST = 'ID danh mục không tồn tại',
  CATEGORY_HAS_PRODUCT = 'Bạn xóa tôi thì những sản phẩm sẽ bị mồ côi',
  PRODUCT_EXIST = 'Sản phẩm đã tồn tại',
  CART_ID_NOT_EXIST = 'ID giỏ hàng không tồn tại',
  CART_ITEM_ID_NOT_EXIST = 'ID sản phẩm trong giỏ hàng không tồn tại',
  CART_EXIST = 'Giỏ hàng đã  tồn tại',
  ADDRESS_ID_NOT_EXIST = 'ID địa chỉ nhận hàng không tồn tại',
  PERMISSION_EXIST = 'Quyền đã tồn tại',
  PERMISSION_ID_NOT_EXIST = 'ID quyền không tồn tại',
  PROMOTION_NOT_EXIST = 'Mã giảm giá không tồn tại',
  PROMOTION_EXIST = 'Mã giảm giá đã tồn tại',
  PROMOTION_INVALID = 'Mã giảm giá không hợp lệ',
  PROMOTION_EXPIRED = 'Mã giảm giá không còn hiệu lực',
  PROMOTION_CONDITIONS_NOT_MET = 'Không đủ điều kiện áp dụng',
  INVALID_DISCOUNT_DATE_RANGE = 'Ngày kết thúc phải lớn hơn ngày bắt đầu',
  INVALID_DISCOUNT_AMOUNT = 'Số tiền giảm giá không hợp lệ',
  INVALID_DISCOUNT_PERCENTAGE = 'Phần trăm giảm giá không hợp lệ',
  COLOR_NOT_EXIST = 'Màu sắc không tồn tại',
  SIZE_NOT_EXIST = 'Kích cỡ không tồn tại',
  OUT_OF_STOCK = 'Sản phẩm không còn đủ hàng',
  ORDER_NOT_EXIST = 'Đơn hàng không tồn tại',
  ORDER_CANNOT_UPDATE = 'Đơn hàng không thể cập nhật',
  ORDER_STATUS_INVALID = 'Trạng thái đơn hàng không hợp lệ',
}